Handle merging of duplicate data sources generated from dbt metrics
different dbt metrics might have the same `.model` reference. Because
we build a data source from each dbt metric, this means we might have
duplicate data sources. Additionally, the dbt metrics don't have to
specify all the dimensions / timestamps that exist on a data source.
Thus we want to gather the dimensions and measures that exist on the
variations of the data source.