Refactor _build_dimensions and helpers for dbt metric transformation
Previously _build_dimensions called helpers to build dimensions from
the different parts of a DbtMetric. However, it turns out there is
additional context needed (like is the filter already listed as a dimension)
such that it makes sense to condense the _build_dimensions and it's
helpers into one _build_dimensions function which is still comprehensible.