Add building data source dimensions from dbt metric filters
Each filter on a dbt metric is composed of 3 parts: the field, the
operator, and the value. Each filter field is a column that must exist
on the data source referenced by the metric's model field. We use the
metric's model to define a data source, and thus each field is a dimension
on that data source