Add agg_time_dimension to measures created from dbt metrics
Use the dbt metric's `.timestamp` value to set the `agg_time_dimension`
for the measure being created on the data source created by the dbt
metric. We must do this because we haven't set a primary time dimension
on the data source. Additonally, even once we begin setting a primary
time dimension for the data source, we want to ensure the measure uses
the time dimension specified by the dbt metric's `.timestamp`