Add dbt-* packages
Note, we're adding these as required dependencies here. However,
we anticiapate marking these as optional and making them extras in
a later commit. Additionally, this commit does not include dbt-databricks
or dbt-duckdb. This is because those packages are still on dbt-core 1.2.x.
They should be upgraded in the near future, and then we'll add them as
well.