Initial refactor of dbt_to_metricflow into class based transformer
In the current implementation we are passing arguments many functions deep,
rerendering nodes multiple time from dbt, and overall have a jumble of code.
This commit is the initial refactor of the code into a more reasonable
class based transformation. As such this commit doesn't reduce the complexity
yet, the following commits will. Having this commit like this though makes
things easier to track.