Refactor time_dimensions_stats to be a property of DbtManifestTransformer
Previously we were having to pass the time_dimension_stats multiple function
levels to where it was used (i.e. it was being passed into functions only
for the purpose to be passed to a lower function). This was cluttery and
bothersome. By lazy loading it as a class instance property we simply this
greatly.