use MetricReference as the accessor key for MetricSemantics
We currently use MetricSpec as a key for everything in MetricSemantics. This restricts us from adding properties to MetricSpec as it would end up messing up the idea of using it as a key on the metric name. To resolve this, replace MetricSpec with MetricReference as the accessor key in MetricSemantics.