Update method for `DbbMetric.model` resolution
Turns out that `DbtMetric.model` can be a dbt `ref` or a simple
model name. At least that is how it looks given some of dbt's test
fixtures (this is not documented). Thus this commit updates the
`DbtMetric.model` resolution method to handle this case and this
commit also updates the documentation for the method slightly