Add validation for validity param specfications
The validity params for defining an SCD window on a data source must
conform to certain specifications, otherwise we will not be able to
properly handle them and users could encounter strange runtime
failures.

This commit adds all of the relevant assertions, including guards
against configurations which we should support, but have not
implemented yet.