Move from dict to default dict tracking data sources created from dbt
Previously we weren't using DefaultDict, but a regular dict for tracking
the data sources created from dbt metrics. This resulted in us having an
if/else statement for checking if the data source name existed in the dict
and then doign some extra work. By using DefaultDict, this is no longer
necessary