Add validation to block DISTINCT keyword on COUNT aggregations
The DISTINCT keyword is a problem for aggregation expressions,
because it effectively overrides the additivity of the aggregation
type. Since we rely on the aggregation type to tell us whether or
not a given measure is additive, we run the risk of doing an
inappropriate partial aggregation -> re-aggregation of these
measures.

This would produce incorrect results for a subset of queries,
which makes the framework behavior unpredictable.

Rather than allowing for this, we block the DISTINCT keyword from
use in COUNT expressions. Follow-ups should enable this safeguard for
SUM and AVG types, but COUNT(DISTINCT x) is generally more common
than SUM(DISTINCT x) so we start here.