Prevent measure definitions in data sources with validity windows
Defining measures in an SCD-style data source is a bit tricky,
because it's unclear when (or if) the user wants the measure to be
globally aggregable vs semi-additive to the validity windows.

While we sort out the right model syntax for this, we will prevent
measure definitions in data sources with validity windows to ensure
there is no mismatch between framework behavior and user expectations
with regard to how these things are aggregated.