Update `get_dbt_project_manifest` to accept `profile` and `target` args
One can have multiple dbt profiles defined, and each dbt profile can
have multiple targets. In most cases the default target for a profile
is the `dev` target, not the `prod` target. Given that there can be
multiple profiles and targets, and that the default profile/target
combination used is likely not the `prod` combination, it's important
to allow the specification of which `profile` and `target` to use as
we want to support prod use cases.