Improve legibility of issues raised by `validate_safely`
The decorator `validate_safely` has been very useful. Specifically
it catches issues during validation that we don't expect to happen.
The problem recently has been that the issues built by `validate_safely`
have become increasibly hard to read. This is mostly because the args
passed to the method which ultimately raised an exception have been
included in the base validation issue message. When the passed arg(s)
are just a data source name, thats not a problem. But when the passed
arg(s) incldue the entire UserConfiguredModel, it becomes an
incomprehensible wall of text. This commit moves the method arguments
and method name to the `extra_details` portion of the validation issue.
Doing so makes the default printing of the validation issue more
legible but still allows to get the detail information if the
validation issue is printed verbosely.