Enable pre-commit typechecking for dbt packages
Metricflow has an override in its .pre-commit-config to ignore all
missing imports, which means third party packages simply don't get
typechecked in any case where the mypy invocation is mediated by
pre-commit (which is roughly all of them, for us).

This means the dbt integration was done with a relative minimum of
typechecker support.

Here we enable dbt typchecking in the .pre-commit-config (the remaining
changes were an unfortunate side effect of auto-formatting) by adding
dbt-core to the dependency list, and clean up the typing errors this
exposed as a result.

Note these fixes are just wallpapering over the underlying issues, so
more robust changes may be warranted.