Run Postgres tests in merge-blocking action
We've had several feature branch merges with tests that passed
in DuckDB but not in other engines. While running all engine tests
on PRs would be impractical, Postgres is perfectly fine - it's
fast, it can spin up a Postgres server within the CI runner itself,
and it covers nearly all of the issues that might pass within DuckDB
but fail in other engines.