Make derived metrics tests use double types
The derived metrics tests were all implicitly checking BIGINT types,
since that's what SUM returns. This changes them to use DOUBLE
for proper ratio computation, which will bring the answers in line
with expectation based on manual calculation against the table fixtures.

Note we cannot do explicit casting in the metric definition expr because
DOUBLE types are not consistently named across engines, so we rely on
the implicit casting shorthand of multiplying by a double literal.