Add `extra_detail` to data warehouse validation errors raised from exceptions
We're having some spec building issues when buiding specs for the
measure based data warehouse validations. It's extra hard to debug
because currently we don't return the exception stacktrace information.
Without this it's hard to see what is going on. This commit adds the
exception stacktrace information to the `extra_details` section of
the validation error. Note, from a CLI perspective `extra_details`
aren't printed out by default. They are only printed when
`--verbose-issues` is specified