Move cumulative metric join creation to  SqlQueryPlanJoinBuilder
The cumulative metric join logic is fairly involved, and as such it
adds considerable clutter up the dataflow to sql visitor module. This
moves it to the join builder class for further consolidation.

This commit is largely a cut and paste of the relevant pieces.
Consolidation of logic will follow separately for ease of tracking.