Add configuration constasts for dbt cloud job_id and service_token
To support running MetricFlow via dbt-cloud projects, we need to hit
the dbt Metadata API for the relevant information. To do MetricFlow
needs two things a `job_id` and a `service_token`. A `job_id`
is necessary because MetricFlow can only work ontop of one dbt job
at a time presently. Additionally the `service_token` is necessary
because the dbt Metadata API is auth gated (as it should be).