Update the MetricFlow engine to build from dbt cloud when detected
The MetricFlow engine has a function `from_config` for building
the engine from defined configs. Specifically this is most typically
used by the `mf` CLI. This commit updates that function to detect
when dbt cloud is meant to be used, and then build the UserConfiguredModel
for the semantic layer for the engine from dbt cloud.