Use SqlQueryPlanJoinBuilder for aggregate measures groupby joins
The aggregate measures join on groupby logic relied on a special
helper for building the join condition, since we need to ensure
all valid measure computation values are included in the INNER JOIN
logic, and FULL OUTER JOIN will not necessarily return the
correct results.

This change eliminates that custom helper in favor of allowing for
a "join on null" behavior via the ColumnEqualityDescription, as
other join types might require this. That change allows us to
use the SqlQueryPlanJoinBuilder.make_sql_join_description
method.

The rest of the join building logic remains embedded in the visitor
because it is particular to the circumstances of a fully aggregated
data set, and so exposing a generic helper did not seem appropriate.