Fix schema for metric inputs to derived metrics
The schema for metric inputs to derived metrics allows for a list
of either string or metric input struct, but the Pydantic object
only accepts this latter. This removes the option from the
schema for the sake of consistency, and also to ensure any editor
plugins based on the jsonschema provide the correct completion
hints.