Rename DbtCloudJobArgs to DbtCloudConfigs and move to CLIContext
The DbtCloudJobArgs were useful where they were because they:
1. Made passing of the args through multiple functions cleaner
2. Made reading/storing the the relevant config values easier

However, having it defined where it was was probematic because
when importing it, the rest of the file would be checked and blow up
because `dbt_metadata_client` is an optional dep. Thus it had to be
moved somewhere. The "best" solution might be to put it in its own
separate file, but that seemed a little extra and there wasn't a great
place for that file itself to live. Instead we moved it to the CLIContext.
This solution still gives us the value of 2, but we lose the value of 1.

Additionally, it's worth noting that the comment the class previously
had about being necessary for use with dbt was inaccurate. This object
isn't used with the dbt core library. That comment was an accidental
copy and paste from `DbtProfileArgs`, and wasn't accurate for
DbtCloudJobArgs.