Add IdentifierType.NATURAL and convert SCD models to use it
The IdentifierType.PRIMARY implies uniqueness, and as such having
SCD-style data sources overload that term can lead to confusion since
the identifier is not unique unless a validity window filter is
applied.

This commit adds a new identifier type - natural, representing the
"natural" key - and updates the scd models to use it. This also
updates the join validator to maintain the same behavior as it had
previously, which means SCD -> SCD join paths will raise exceptions
whenever they are encountered, instead of being filtered out.