Add NULL valued dimensions to integration test model tables
Our model datasets do not contain any dimension rows with null
values for dimensions of interest. This adds an "incomplete" entry
to the listings tables used by the simple model and the scd model
so that our tests will typically cover scenarios where a dimension
value might be NULL.

This fixes the improper JOIN configuration in the dimension tests,
as those began failing. The framworke uses LEFT OUTER JOIN for
dimensional joins, and since the user associated with the missing
listing does not exist in the users tables the use of the wrong
JOIN type was filtering out the NULL valued rows.