Move to `convert` and `map` terminology from `transform`
Transformation terminology is commonly associated with the process of
transformation rules which take in a user configured model, and return
a MetricFlow model. Thus using `transfrom` was highly confusing in the
context of converting dbt metrics into a MetricFlow model. The individual
rules themselves we've moved to calling them `mapping_rules`. This is because
the rules litterally map one attribute in a dbt metric to it's MetricFlow
model counterpart. The `DbtTranformer` has become the `DbtCoverter` and
`converts` instead of `transforms` because that better communicates what
is actually happening.

Additionally, we've started using (at a variable and funciton  naming level)
`metricflow_model` instead of just `model`. That is because we want to start
building on the idea that a `UserConfiguredModel` and a `MetricFlowModel` are
different types. However, the `MetricFlowModel` class hasn't been actually
created yet.

Resolves: https://github.com/transform-data/metricflow/pull/324#discussion_r1031850184