Add test scenario for duplicate rows in full outer joins
When we combine metric nodes we do FULL OUTER JOINs across
all queried dimensions to link them together. While FULL OUTER
JOIN is commutative for two inputs, this is not true when three or
more inputs are FULL OUTER JOINed together. In particular, multiple
rows with the same sets of dimension values may be emitted if we
do not make an ON condition that coalesces all previously seen
values from a given input row on the left of the join step. See
the updated documentation on CreateOnConditionForCombiningMetrics
for an example.

This commit updates test data to trigger this condition for the
three-metric integration test.