Begin using custom dbt metadata api metrics query
Previously we were using the `get_metrics()` method provided by the
dbt_metadata_api package's `Client`. However this only goes two levels
deep for the graphql API call. That is for a given metric, it gets attributes
and nodes directly part of a metric node. However any sub nodes were only
getting associated attributes, not associated sub nodes. This was fine.
However we now, with the next commit, going to need to get the columns
associated with a model associated with a metric. Because the columns are
sub nodes (CatalogColumns), we now need to specify them. Due to how
`dbt_metadata_api` builds query operations, we need to build out the full
query of what we need.