Update `DbtDimensionsToDimensions` rule to associate dimension types with columns
Preivously in the rule `DbtDimensionsToDimensions` we marked all dimensions
created by the rule as `CATEGORICAL`. We did this because we new that
if a dimension showed up in a dbt metric filter or the dbt metric timestamp
it would be updated to be a timestamp. However, this had two problems. First,
it required that the mapping rules be run in a specific order, which we
don't want to guarantee. Secondly, it mean that if a dimension should be
`TIME` but didn't get corrected by the filters or timestamp rule, it'd be
incorrectly typed as `CATEGORICAL`. This commit fixes both of those issues.