Add ubuntu version configuration to the cache key
The CI runner was failing to find matching versioned dependencies in
the hosted tools cache. This was likely caused by one of two things:

1. An update to Ubuntu that updated some system packages
2. A mis-configured cache key

The former is fixed by appending the ubuntu version to the cache key
in our actions/cache configurations.

The latter is fixed by providing the hashFiles call in our actions/cache
configuration with the correct file path for the config file we wish to
hash.

This commit does both of these things, because they're so closely
related and various CI runs are failing for one or the other reason.