Enable `DbtFiltersToDimensions` to parse dimensions from `filter.value`
It was recently discovered that one can absolutely put a dimension
as the value for a filter. It wasn't documented, but in testing it was
found to be a valid dbt config. This update adds that context to the
`DbtFiltersToDimensions` rule. We could probably have put this in its
own rule, but after deliberating with myself for 30 minutes I decided
to just have it in `DbtFiltersToDimensions`. Splitting hairs here.