Simplify `DbtToMetricConstraint` to just build the SQL constraint clause
We were going the extra step previously of building a WhereClauseConstraint
object for each dbt metric filter. This is more work than we really need
to do. Instead now we just build the SQL statement given the dbt metric
filters, and when the mapped objects are parsed to MetricFlow Model elements,
the WhereClauseConstraint parser handles building the constraint objects.
The primary reason for this change was that we now know that dbt filter values
can be dimensions. If we kept doing things how we were for this rule, this
would mean having to detect when the filter value was a dimension. Instead
of adding that complexity, we decided to simplify.