Consolidate metric output rows with null dimension values
The production behavior of the CombineMetricsNode is to run a
FULL OUTER JOIN when multiple metrics are requested in output. The
FULL OUTER JOIN itself does a set of equality checks, and returns
a match if all linkable values (dimensions) pass an equality check.

If any dimension has a NULL value, we essentially compare NULL = NULL,
which returns NULL, and fails the test in the ON condition. This results
in both rows being included independently, which results in duplicate
entries for the given tuple of dimension values.

This PR adds an aggregate function to the select statement for the
cases where we use a FULL OUTER JOIN with the combine metrics node, which
effectively deduplicates these artificially duplicated records.

Note a COALESCE or direct NULL comparison would be preferable, but the
former requires casting to a common type, which Redshift does not support
without custom rendering for boolean values (which is not currently possible
in MetricFlow), and the latter does not work with FULL OUTER JOIN types in
most of the engines we support.