Build combine metric joins based on from_data_set and join_data_sets
The dataflow to sql converter for the combine metrics node relied on
an index walk across the list of all input datasets, treating the
zero-th entry as "from" and building a set of comparison aliases
for the coalesce expressions based on the entire sequence on
every step through the loop.

This is a bit convoluted, and makes the join iteration inconsistent
with the way we would do it when building an INNER JOIN or CROSS JOIN
across these datasets. In preparation for adding handling for those
scenarios, this commit does a bit of renaming and restructuring to
ensure that we are producing the same results in the FULL OUTER JOIN
case, but doing so while iterating through the join data sets instead
of a range bound sequence.