Fix incomplete JOIN condition for combining derived metrics nodes
The CombineMetricsNode was originally designed to join together multiple
fully computed metrics nodes for rendering unified reseults to an
end user. As such, it relied on a FULL OUTER JOIN, and doing so in
a way that failed to account for NULL values on the selected dimensions
would result in potential duplicate sets of dimension values.

In addressing this annoyance we noticed that derived metrics were using
this node to join input metrics for subsequent expression computation,
and using an INNER JOIN for consistency with multiple-measure
computations, which themselves use INNER JOIN because we have no
current mechanism for users to consistently handle NULL inputs to
measure/metric expressions.

This means the on condition MUST account for NULL inputs correctly. The
original join, which was essentially a.dimension = b.dimension, would
exclude the joined row altogether if both a.dimension and b.dimension were
NULL. The technique we were forced to adopt for the FULL OUTER JOIN case
would not work here, because that does a re-aggregation pass, but here we
need to prevent the rows from being thrown away.

Therefore, this change adds an additional set of IS NULL checks and passes
the row if both inputs are NULL, using the exact same mechanism we used
for the INNER JOIN on multi-measure inputs.