Move combine metrics join building logic into DataflowToSqlQueryPlanConverter
The original join logic for the combine metrics node was encapsulated in
an InstanceSpecTransform which would iterate through all of the instances,
generate column names, and build coalesced equality comparisons for
the on condition in that join.

This was a sensible approach when this function would only ever handle
one singular join type, but now that its scope has expanded to two join
types that build different on statements with that same list of columns it
makes more sense to consolidate the logic in the SqlQueryPlanJoinBuilder.

This commit is an interim step to remove the original InstanceSpecTransform
subclass and inline the logic in the DataflowToSqlQueryPlanConverter. The
next commit will move all of this logic into the SqlQueryPlanJoinBuilder.