Update `DbtFiltersToDimensions` to handle complex filters
Filters for dbt metrics can really be anything. This is because dbt
filter fields, operators, and values are concatinated to make an sql
clause. The best way for us to make sure were detecting all dimensions
in a filter is to actually construct the clause and then parse it with
MetricFlow's WhereClauseConstraint. This is what we now do.