Add all group by columns to select in semi-additive join node
Under certain conditions the semi-additive join node rendering
produces slightly strange SQL, where columns in the GROUP BY
statement are not included in the SELECT. For example, the inner
query for the filter join might look like this:

```
SELECT MIN(ds) AS ds__complete
FROM input_table
GROUP BY DATE_TRUNC(ds, isoweek)
```

This is confusing, as the DATE_TRUNC expression in the group by
allows for mulitple rows of output. In this case, since the `ds`
columns are dates, there will always be exactly 1 row per isoweek,
however it's not obvious that this is the intent of the query.

This also causes problems with certain rendering approaches, as
we typically render either a raw expression OR an alias in GROUP BY
statements. In cases where we render the latter, this query will
break because the DATE_TRUNC() *must* be rendered as a full expression.
Depending on inputs and optimizer layouts this could, for certain engines,
result in broken SQL, a problem which is easily avoided if the expression
is included in the SELECT statement as a properly aliased column expression.

This change adds the full set of GROUP BY columns to the SELECT expression,
which results in the following updated SQL (for expr-style group by references):

```
SELECT DATE_TRUNC(ds, isoweek) AS ds__week, MIN(ds) AS ds__complete
FROM input_table
GROUP BY DATE_TRUNC(ds, isoweek)
```