Stop updating enum values by reference in BigQuerySqlExpressionRenderer
The BigQuerySqlExpressionRenderer is updating the `value` property of
an enum value. It's possible Python passes enumeration types by value,
or does some copy on write, but most likely they're passed by reference
like everything else. As such, modifying properties on an enum instance
is likely bad form.

The existing logic also blocks mypy updates to 0.940 or later. So we
fix it here.