Stop returning the model from ModelValidator
Model validation rules do not alter the model. By extension the
ModelValidator does not alter the model. The fact that the ModelValidator
was returning the model was a carry over from a now long bygone era
when model validators and model transformations were much more tightly
coupled. We've, for the better, gotten away from that. This change cleans
up this no longer needed functionality. Indeed it's desireable to stop
returning the model from the ModelValidator because we want to cememnt the
idea that the model validator does not and should not modify the model.