Restrict pre-commit to versions in [2.18.0, 3)
Pre-commit recently released a 3.0, but our project configuration is
restricted to version 2.x (specifically, 2.18 or higher, but less than
3.0).

This is causing failures in CI, as the command in this config allows
for installation of the latest pre-commit, and certain changes in
pre-commit 3.0 are causing problems with our existing configuration.

Since these updates are more involved than a simple version change,
we limit the CI version range for the time being.