Remove `owners` attribute from test fixtures semantic layer objects
Previously we were accidentally using `schemas_internal` in some areas of
MetricFlow. This was problemantic for many reasons... In the cutover we
corrected this by ensuring parsing uses `schemas` instead of `schemas_internal`.
This in turn broke SO MANY TESTS. I intended to start making changes to the tests
once we moved from cutover stuff to modifications. But turns out I needed
to do some modifications during the cutover. Thus this change.