Add `descriptions` to jsonschema object definitions in `schemas.py`
We weren't planning on making changes to things during the cut over,
but it turns out in some instances we need to. We plan on having
descriptions on the semantic layer objects, and the test fixture
semantic layer objects currently define descriptions. Thus instead of
removing descriptions from the tests and re-adding them when we start
to update the schemas, we're instead saying it's saner to just add
`description` to the object defintions now.