Remove use of `SqlTable` in `reserved_keywords.py` validation
Validations are going to be moved over to dbt_semantic_interfaces.
As such they can't be importing metricflow. The change here is small.
We're unfortunately duplicating code that is in `SqlTable`, but this
will likely be changing again in the near future as dbt_semantic_interfaces
plans on using a more structured object directly on the data source
in place of sql_table.