Remove `SUPPORTED_GRANULARITIES` check from `DataSourceTimeDimensionWarningsRule`
These validations are moving to dbt_semantic_interfaces. In dbt-semantic-interfaces
there won't be a defined `SUPPORTED_GRANULARITIES` just all the TimeGranularty
Enum, and checking against this would be pointless because it has to be
a valid TimeGranularity before this point. Long term if MetricFlow for
some reason doesn't support a granularity defined in dbt-semantic-interfaces,
then at that point we should write a custom rule in MetricFlow validating
that only the supported granularities are used.