Move dataclass_serialization.py to dbt_semantic_interfaces
We're not sure if this serialization is actually still necessary, it
may be vestigial. However, until we figure that out, dbt-semantic-interfaces
will need to know about SerializableDataclass. This is because one of the
main things that can be serialized are instances sets. Although instance
sets aren't in dbt-semantic-interfaces, references are, and instance have
a lot of references. Thus the references must be built with
SerializableDataclass, and thus it must be in dbt-semantic-interfaces.
If SerializableDataclass does turn out to be vestigial, then we'll remove
it from things at that point.