Making AggregationType use ExtendedEnum again
When we moved AggregationType over, we changed it from an ExtendedEnum
to an Enum. We thought this wouldn't be problematic. However, it turns out
that the `_missing_` function of ExtendedEnum is actually very important.
This is because the custom `_missing_` implementation ensures that using
uppercased or lowercased values in a config file both work. That is
to say that with `ExtendedEnum`, `sum` and `SUM` are both valid, but
with `Enum` only `sum` is.