Move TimeGranularity to new type_enums directory
TimeGranularity had been living in `objects`. However `objects` is becoming
`implementations`, that is implementations of protocols. `TimeGranularity`
isn't an implementation of a protocol, nor is it a protocol. Additionally
it is a required  typing on our protocols. Thus it made sense to create
a new directory, `type_enums`, to hold `TimeGranularity` as well as our
other extended enums used to restrict typing.