Add tests asserting that our implementations are satisfy the protocol defintions
Note: These tests are in MetricFlow. That is because we haven't moved tests
over to dbt_semantic_interfaces yet. Having this test in dbt_semantic_interfaces
dir would require us to update the make file for running tests as well as
the github workflows. And that is not desireable at this time.