Re-enable test for lonely entities and remove test's dependency on MF specs
We disabled awhile ago because we had disabled the validation. We have since
re-enabled the validation (a while ago at this point), thus it makes
sense to re-enable the test. Additionally the test doesn't need to rely
on MF specs, and moreso CANNOT depend on MF specs. Thus we've removed
that dependency.