Switch mypy to use the local environment for installed packages
Our initial approach to using mypy was to have a .pre-commit-config
with a list of additional dependencies needed by mypy to do typechecking
on third-party packages. These dependencies cannot be easily updated in
lock-step with the package configuration we use in poetry, an issue which
has caused our CI to break in the past due to version mismatches.

Since all of our CI should build the package from a fresh virtual environment,
we have decided to risk the discrepancies that might arise for local
developers in order to ensure mypy is running against the same build
state (modulo a handful of typing packages and optional "dev-only"
installations of optional packages).

This commit moves to local evnironment dependencies and updates the mypy and
CI linting configurations to account for the change.