Skip following imports in ruamel.yaml
Every time we change anything mypy blows up with spurious ruamel.yaml
typechecking errors, claiming that it cannot find the type stubs. This
is aggravating. In fact, sometimes it blows up when something else
changes and then we need to run poetry update for no apparent reason.

We'll try skipping ruamel.yaml here in the vain hope that this will
make this whole mypy + pre-commit house of cards less fragile.

Also, apologies for the "while I'm here" update to mypy.ini,
but the fuzzywuzzy override is no longer needed as it is not part
of the package anymore.