Cut over MetricFlow dependencies to dbt-semantic-interfaces deployment
We have reached the stage where it is time to move MetricFlow's internal
dependencies on our SemanticModel types over to the definitions in
dbt-semantic-interfaces.

In order to minimize the number of inline changes needed to keep the packages
aligned, this commit points MetricFlow to an initial dev build without the
full SemanticModel protocols in place and only updates the handful of places
where path locations have drifted out of sync.