Fix type signatures for log_call decorator
The original implementation of the log_call decorator returned an
object of type Callable[..., Any] because the version of mypy we
were using at the time did not support parameter binding correctly.

This resulted in implicit Any typing for every method wrapped with
this decorator, which means mypy would not catch errors with input
parameters. For example, the MetricFlowEngine.query() method is
typed to accept a single parameter, but callers could pass in any
combination of parameters without a typechecking error because
the `query` method was wrapped in a decorator that converted it from
it's original signature to Callable[..., Any].

Now that we are up to date on mypy, we can follow the typing guide
for decorators (or, as in this case, decorator factories) to
ensure the wrapped function types remain consistent.