Roll pre-release version number back to 0.200.0.dev2 from 0.200.1.dev1
Somehow or other we ended up in 0.200.1. This rolls us back to
0.200.0 as our base version, and pushes the dev build up to the next
version mark.

The downside of this is 0.200.1.dev1 is no longer a usable version
number, but we can contend with that if we come to it. Given
MetricFlow's current state of development it's unlikely we'll need
to roll multiple dev builds on patch releases, so this seems like
the right trade.