Update legacy dbt dependency spec to cover future requirements
Our legacy dbt integration laid out a bunch of optional dependencies
on dbt core 1.4 plus various adapter packages.

The integration plan for MetricFlow 0.200.0 involves a dependency
link with dbt core 1.6 in order to enable query execution through
dbt adapters and reading from dbt project metadata for parsing
the semantic manifest.

This commit removes the legacy dependency structures and adds a
temporary dependency on dbt core 1.5. For the purposes of our
current integration steps 1.5 is adequate, and at the moment the
1.6 beta releases do not appear to be published for all of the
adapters we want to test.