Merge AsyncSqlClient with SqlClient
The division between the AsyncSqlClient and SqlClient exists for
historical reasons which no longer apply. We are now moving away
from the async client with its own background thread in favor
of a more streamlined SqlClient interface. As a first step,
we merge the separate protocols into a single interface, which will
make it easier to modify the interface itself without needing to
juggle the SqlClient/AsyncSqlClient type duality.