Add support for query tags and isolation in main query and execute methods
Previously, support for query tags and isolation level were only available
in the AsyncSqlClient protocol. As part of the streamlining effort we will
be consolidating the async_query and query methods, so we add support
for these features to the base query methods in order to be able to
remove the async variations.