Remove query cancel methods from sql client
The current implementation of query cancellation relies heavily on
the async query interfaces, which we are removing in favor of
making MetricFlow a single threaded single query execution library.
We may need to bring the cancel method back in some form, but it'll
be more similar to what we see in libraries like the dbt adapter
library, where each client implements a "cancel all queries for
a given connection" kind of interface.

This commit removes the cancel method to streamline the process of
removing the async query and async execute commands.