Move health_check functionality into CLI package
For legacy reasons having to do with Snowflake-specific health check
requirements for the Transform product, the health check command in
the CLI delegated all query execution into the SqlClient class associated
with the warehouse engine.

We no longer have any need for engine-specific customization, and so
this method can be moved out of the SqlClient protocol and into the
CLI health check method.